authorization implemented

+++ src/api/notes/handler.js
@@ -1,11 +1,9 @@
+/* eslint-disable no-underscore-dangle */
 const ClientError = require('../../exceptions/ClientError');
 
-/* eslint-disable no-underscore-dangle */
 class NotesHandler {
   constructor(service, validator) {
-    // eslint-disable-next-line no-underscore-dangle
     this._service = service;
-    // eslint-disable-next-line no-underscore-dangle
     this._validator = validator;
 
     this.postNoteHandler = this.postNoteHandler.bind(this);
@@ -19,9 +17,11 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      // eslint-disable-next-line no-underscore-dangle
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -53,9 +53,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    // eslint-disable-next-line no-underscore-dangle
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -67,7 +67,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
-      // eslint-disable-next-line no-underscore-dangle
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -98,10 +100,10 @@ class NotesHandler {
 
   async putNoteByIdHandler(request, h) {
     try {
-      const { id } = request.params;
       this._validator.validateNotePayload(request.payload);
+      const { id } = request.params; const { id: credentialId } = request.auth.credentials;
 
-      // eslint-disable-next-line no-underscore-dangle
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
@@ -132,8 +134,11 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
-      // eslint-disable-next-line no-underscore-dangle
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
+
       return {
         status: 'success',
         message: 'Catatan berhasil dihapus',
